backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
pymongo==4.9.1
pydantic==2.9.2
pydantic-settings==2.5.2
PyJWT==2.9.0
cloudinary==1.41.0
python-multipart==0.0.12
python-dotenv==1.0.1
bcrypt==4.2.0
email-validator==2.2.0
